requirements.txt
Pillow==10.1.0
bcrypt==4.0.1